Create README.md
A game I made to familiarize myself with javascript DOM manipulation. A user is given a shape of a particular color to select. The color of the text for the color and shape provided to the user may be of the colors of the other shapes available to select, which makes guessing the correct answer more difficult. Scoring is based on how many shapes are selected in the allotted time.


The game was based on an RGB color picker game made by Colte Steele  (https://github.com/Colt?tab=repositories)